refactor : Change API project's entire structure
