+++ docker-compose.yaml
@@ -1,0 +1,26 @@
+version: '3.7'
+services:
+  api:
+    build: .
+    image: "capston-design-be"
+    ports:
+      - "3000:3000"
+    restart: "unless-stopped"
+    depends_on:
+      - db
+    networks:
+      - api-network
+  db:
+    image: mysql
+    ports:
+      - "3306:3306"
+    restart: "unless-stopped"
+    environment:
+      - MYSQL_ROOT_PASSWORD=hoplin1234!
+      - MYSQL_ROOT_HOST=%
+      - MYSQL_DATABASE=cloudedu
+    networks:
+      - api-network
+networks:
+  api-network:
+    driver: bridge
