+++ nest-cli.json
@@ -6,7 +6,7 @@
     "deleteOutDir": true,
     "assets": [
       {
-        "include": "./config/**/*.env",
+        "include": "./**/*.env",
         "outDir": "./dist"
       }
     ]
