+++ src/app/health/types/ping.response.ts
