+++ src/infrastructure/abstract-entities/member.interface.ts
@@ -1,5 +1,4 @@
 export interface MemberInterface {
-  id: string;
   email: string;
   name: string;
   password: string;
