+++ src/infrastructure/entities/department.interface.ts
