+++ src/infrastructure/types/index.d.ts
@@ -7,3 +7,5 @@ enum Sex {
   MALE = 'male',
   FEMALE = 'female',
 }
+
+type FieldOptional<T, K extends keyof T> = Omit<T, K> & Partial<T>;
