--- src/member/dto/create-member.dto.ts
@@ -1,1 +1,0 @@
-export class CreateMemberDto {}
