--- src/member/dto/update-member.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { CreateMemberDto } from './create-member.dto';
-
-export class UpdateMemberDto extends PartialType(CreateMemberDto) {}
