--- src/member/member.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { MemberController } from './member.controller';
-import { MemberService } from './member.service';
-
-describe('MemberController', () => {
-  let controller: MemberController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [MemberController],
-      providers: [MemberService],
-    }).compile();
-
-    controller = module.get<MemberController>(MemberController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
