--- src/member/member.service.ts
@@ -1,26 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { CreateMemberDto } from './dto/create-member.dto';
-import { UpdateMemberDto } from './dto/update-member.dto';
-
-@Injectable()
-export class MemberService {
-  create(createMemberDto: CreateMemberDto) {
-    return 'This action adds a new member';
-  }
-
-  findAll() {
-    return `This action returns all member`;
-  }
-
-  findOne(id: number) {
-    return `This action returns a #${id} member`;
-  }
-
-  update(id: number, updateMemberDto: UpdateMemberDto) {
-    return `This action updates a #${id} member`;
-  }
-
-  remove(id: number) {
-    return `This action removes a #${id} member`;
-  }
-}
